chore(composer): bind to latest 2.0 branch contents
The Composer docs for selecting a branch are wrong (suggesting `dev-2.0` when
packagist actually makes this branch available as `2.0.x-dev`). While we're
testing Elgg 2.0 I think it's annoying to require a particular beta. Better
to let testers install the latest and be able to simply `composer update` to
get the latest fixes.